Set or override config properties via cli arguments
This allows to accomodate customizing deliveries for different customers
without having to create new manifests, if the differences between them
are not in the chapters. I.e. the difference is only the name of the
customer.

So now if we have the same beginning course for company Foo and Bar
we are still able to customize things like the pdf, archive and watermark
without having to create a new config.yaml.